agents/tools/tools-applications/rag-as-a-tool/rag-patterns/RAG/naive-rag/rag-faiss/naive-rag-faiss-poc/requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
openai
python-dotenv